% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Funcao para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{um numero}

\item{b}{outro numero}
}
\description{
Esta funcao serve para somar dois numeros
}
\examples{
somar(3,4)

}
